% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_excel_with_dictionary}
\alias{read_excel_with_dictionary}
\title{Create a coded and labelled data frame from an Excel File}
\usage{
read_excel_with_dictionary(
  data_file,
  data_sheet,
  dictionary_sheet,
  dates_as_character = FALSE,
  na = ""
)
}
\arguments{
\item{data_file}{The path to the Excel data to be imported}

\item{data_sheet}{The name of the sheet to import (required)}

\item{dictionary_sheet}{The name of the shet with the data dictionary. This
must be formatted by the PMHbiostats macro file}

\item{dates_as_character}{will format the dates in unambiguous yyyy-mm-dd
format and return character variables. This can be useful for identifying
date problems. The default is FALSE, to return date-type variables.}

\item{na}{Character vector of strings to interpret as missing values. This
will be applied to the dictionary as well as the data}
}
\description{
Creates a cleaned, labelled and coded data set from data that the excel data
checker has been run on
}
\details{
The variable labels from the dictionary are added, factors are created from
coded variables and date variables are converted to character representation.

The following columns are expected to appear in the dictionary:
Current_Variable_Name,	Suggested_Name,	Label_For_Report,	Type,	Value,
Value_Label, Column_Number,	Import (although Suggested_Name may be missing if
the variable names were already clean)
}
